+++ tests/TestCase/Command/ExportCommandTest.php
@@ -38,8 +38,8 @@ class ExportCommandTest extends TestCase
         $this->exec($this->command);
         $this->assertExitWithSuccess();
         $this->assertOutputContains('Connection: test');
-        $this->assertOutputContains('Driver: Mysql');
-        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_test_\d+\.sql` has been exported/');
+        $this->assertOutputRegExp('/Driver: (Mysql|Postgres|Sqlite)/');
+        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql` has been exported/');
     }
 
     /**
@@ -50,7 +50,7 @@ class ExportCommandTest extends TestCase
     {
         $this->exec($this->command . ' --compression bzip2');
         $this->assertExitWithSuccess();
-        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_test_\d+\.sql\.bz2` has been exported/');
+        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql\.bz2` has been exported/');
     }
 
     /**
@@ -73,7 +73,7 @@ class ExportCommandTest extends TestCase
         $files = $this->createSomeBackups();
         $this->exec($this->command . ' --rotate 3 -v');
         $this->assertExitWithSuccess();
-        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_test_\d+\.sql` has been exported/');
+        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql` has been exported/');
         $this->assertOutputContains('Backup `' . basename(array_value_first($files)) . '` has been deleted');
         $this->assertOutputContains('<success>Deleted backup files: 1</success>');
     }
@@ -86,8 +86,8 @@ class ExportCommandTest extends TestCase
     {
         $this->exec($this->command . ' --send mymail@example.com');
         $this->assertExitWithSuccess();
-        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_test_\d+\.sql` has been exported/');
-        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_test_\d+\.sql` was sent via mail/');
+        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql` has been exported/');
+        $this->assertOutputRegExp('/Backup `[\w\-\/\:\\\\]+backup_[\w_]+\.sql` was sent via mail/');
     }
 
     /**