+++ tests/TestCase/Command/ImportCommandTest.php
@@ -39,7 +39,7 @@ class ImportCommandTest extends TestCase
         $this->exec($this->command . ' ' . $backup);
         $this->assertExitWithSuccess();
         $this->assertOutputContains('Connection: test');
-        $this->assertOutputContains('Driver: Mysql');
+        $this->assertOutputRegExp('/Driver: (Mysql|Postgres|Sqlite)/');
         $this->assertOutputContains('<success>Backup `' . $backup . '` has been imported</success>');
     }