+++ tests/TestCase/Driver/DriverTest.php
@@ -15,13 +15,17 @@ declare(strict_types=1);
 
 namespace DatabaseBackup\Test\TestCase\Driver;
 
-use Cake\Database\Connection;
-use DatabaseBackup\Driver\Mysql;
+use DatabaseBackup\Driver\Driver;
 use DatabaseBackup\TestSuite\TestCase;
 use ErrorException;
+use Symfony\Component\Process\Process;
 
 /**
- * DriverTest class
+ * DriverTest class.
+ *
+ * Performs tests that are valid for each driver class, thus covering the
+ *  methods of the abstract `Driver` class.
+ * @covers \DatabaseBackup\Driver\Driver
  */
 class DriverTest extends TestCase
 {
@@ -32,6 +36,37 @@ class DriverTest extends TestCase
     protected $Driver;
 
     /**
+     * Internal method to get a mock for `Driver` abstract class
+     * @param array $mockedMethods Mocked methods
+     * @return \DatabaseBackup\Driver\Driver&\PHPUnit\Framework\MockObject\MockObject
+     */
+    protected function getMockForAbstractDriver(array $mockedMethods = []): Driver
+    {
+        /** @var \Cake\Database\Connection $connection */
+        $connection = $this->getConnection('test');
+
+        return @$this->getMockForAbstractClass(Driver::class, [$connection], '', true, true, true, $mockedMethods);
+    }
+
+    /**
+     * Internal method to get a mock for `Process` class, that returns a failure
+     *  with a custom error message
+     * @param string $error Custom error message
+     * @return \Symfony\Component\Process\Process&\PHPUnit\Framework\MockObject\MockObject
+     */
+    protected function getMockForProcessWithError(string $error): Process
+    {
+        $process = @$this->getMockBuilder(Process::class)
+            ->setMethods(['getErrorOutput', 'isSuccessful'])
+            ->setConstructorArgs([[]])
+            ->getMock();
+        $process->method('getErrorOutput')->will($this->returnValue($error . PHP_EOL));
+        $process->method('isSuccessful')->will($this->returnValue(false));
+
+        return $process;
+    }
+
+    /**
      * Called before every test method
      * @return void
      */
@@ -39,36 +74,22 @@ class DriverTest extends TestCase
     {
         parent::setUp();
 
-        if (!$this->Driver) {
-            /** @var \Cake\Database\Connection $connection */
-            $connection = $this->getConnection('test');
-            $this->Driver = new Mysql($connection);
-        }
-    }
-
-    /**
-     * Test for `__construct()` method
-     * @return void
-     * @test
-     */
-    public function testConstruct(): void
-    {
-        $this->assertInstanceof(Connection::class, $this->getProperty($this->Driver, 'connection'));
+        $this->Driver = $this->Driver ?: $this->getMockForAbstractDriver();
     }
 
     /**
      * Test for `export()` method on failure
      * @return void
-     * @since 2.6.2
      * @test
      */
     public function testExportOnFailure(): void
     {
-        $this->expectException(ErrorException::class);
-        $this->expectExceptionMessageMatches('/^Export failed with exit code `\d`$/');
-        $config = ['database' => 'noExisting'] + $this->Driver->getConfig();
-        $this->setProperty($this->Driver, 'connection', new Connection($config));
-        $this->Driver->export($this->getAbsolutePath('example.sql'));
+        $expectedError = 'mysqldump: Got error: 1044: "Access denied for user \'root\'@\'localhost\' to database \'noExisting\'" when selecting the database';
+
+        $driver = $this->getMockForAbstractDriver(['_exec']);
+        $driver->method('_exec')->will($this->returnValue($this->getMockForProcessWithError($expectedError . PHP_EOL)));
+        $this->expectExceptionMessage('Export failed with error message: `' . $expectedError . '`');
+        $driver->export($this->getAbsolutePath('example.sql'));
     }
 
     /**
@@ -79,11 +100,9 @@ class DriverTest extends TestCase
      */
     public function testExportStoppedByBeforeExport(): void
     {
-        $backup = $this->getAbsolutePath('example.sql');
-        $Driver = $this->getMockForDriver(Mysql::class, ['beforeExport']);
+        $Driver = $this->getMockForAbstractDriver(['beforeExport']);
         $Driver->method('beforeExport')->will($this->returnValue(false));
-        $this->assertFalse($Driver->export($backup));
-        $this->assertFileDoesNotExist($backup);
+        $this->assertFalse($Driver->export($this->getAbsolutePath('example.sql')));
     }
 
     /**
@@ -92,12 +111,13 @@ class DriverTest extends TestCase
      */
     public function testGetBinary(): void
     {
-        $this->assertEquals(which('mysql'), $this->invokeMethod($this->Driver, 'getBinary', ['mysql']));
+        $binary = DATABASE_BACKUP_DRIVER == 'mysql' ? 'mysql' : (DATABASE_BACKUP_DRIVER == 'postgres' ? 'pg_dump' : 'sqlite3');
+        $this->assertEquals(which($binary), $this->Driver->getBinary($binary));
 
         //With a binary not available
         $this->expectException(ErrorException::class);
         $this->expectExceptionMessage('Binary for `noExisting` could not be found. You have to set its path manually');
-        $this->invokeMethod($this->Driver, 'getBinary', ['noExisting']);
+        $this->Driver->getBinary('noExisting');
     }
 
     /**
@@ -115,18 +135,18 @@ class DriverTest extends TestCase
     /**
      * Test for `import()` method on failure
      * @return void
-     * @since 2.6.2
      * @test
      */
     public function testImportOnFailure(): void
     {
-        $this->expectException(ErrorException::class);
-        $this->expectExceptionMessageMatches('/^Import failed with exit code `\d`$/');
-        $backup = $this->getAbsolutePath('example.sql');
-        $this->Driver->export($backup);
-        $config = ['database' => 'noExisting'] + $this->Driver->getConfig();
-        $this->setProperty($this->Driver, 'connection', new Connection($config));
-        $this->Driver->import($backup);
+        $expectedError = 'ERROR 1044 (42000): Access denied for user \'root\'@\'localhost\' to database \'noExisting\'';
+
+        $driver = $this->getMockForAbstractDriver(['_exec']);
+        $driver->method('_exec')->will($this->returnValue($this->getMockForProcessWithError($expectedError . PHP_EOL)));
+
+        $this->expectException(\ErrorException::class);
+        $this->expectExceptionMessage('Import failed with error message: `' . $expectedError . '`');
+        $driver->import($this->getAbsolutePath('example.sql'));
     }
 
     /**
@@ -137,10 +157,8 @@ class DriverTest extends TestCase
      */
     public function testImportStoppedByBeforeExport(): void
     {
-        $backup = $this->getAbsolutePath('example.sql');
-        $Driver = $this->getMockForDriver(Mysql::class, ['beforeImport']);
+        $Driver = $this->getMockForAbstractDriver(['beforeImport']);
         $Driver->method('beforeImport')->will($this->returnValue(false));
-        $this->assertTrue($Driver->export($backup));
-        $this->assertFalse($Driver->import($backup));
+        $this->assertFalse($Driver->import($this->getAbsolutePath('example.sql')));
     }
 }