+++ tests/TestCase/Utility/BackupImportTest.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
 namespace DatabaseBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
-use DatabaseBackup\Driver\Mysql;
+use DatabaseBackup\Driver\Driver;
 use DatabaseBackup\TestSuite\TestCase;
 use DatabaseBackup\Utility\BackupExport;
 use DatabaseBackup\Utility\BackupImport;
@@ -56,7 +56,7 @@ class BackupImportTest extends TestCase
      */
     public function testConstruct(): void
     {
-        $this->assertInstanceof(Mysql::class, $this->getProperty($this->BackupImport, 'driver'));
+        $this->assertInstanceof(Driver::class, $this->getProperty($this->BackupImport, 'driver'));
         $this->assertNull($this->getProperty($this->BackupImport, 'filename'));
     }